refactor(test): split and clean up test files
