--- pkg/gotest/helpers_test.go
@@ -1,1 +1,0 @@
-package gotest_test
